feat(api): complete Wave 0 contract and validation
Freeze enums and upload contract, isolate SQLite tests, and document API handoff on the personal branch.

CR: CR-20260906-001
Task: T-01, T-02, T-03
Worktree: WT-01

+++ apps/api/src/greenlogix_api/ingest_xlsx.py
@@ -5,6 +5,7 @@ from __future__ import annotations
 import logging
 from datetime import datetime, time
 from io import BytesIO
+
 from openpyxl import load_workbook
 
 from greenlogix_api.models import Order
@@ -58,7 +59,7 @@ def normalize_window(value: object) -> str:
     if isinstance(value, bool):
         return ""
     if isinstance(value, (int, float)):
-        minutes = int(round(float(value) * 24 * 60)) % (24 * 60)
+        minutes = round(float(value) * 24 * 60) % (24 * 60)
         hours, mins = divmod(minutes, 60)
         return f"{hours:02d}:{mins:02d}"
     text = str(value).strip()
